+++ tests/testthat/testTranslations/R/hello.r
@@ -1,0 +1,3 @@
+hello <- function() {
+  gettext("Hello", domain = "R-testTranslations")
+}