+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgload (development version)
 
+* `load_all()` automatically registers package translations, if found.
+
 # pkgload 1.2.3
 
 * pkgload now forces all bindings on unload. This fixes errors and inconsistencies when dangling references force lazy bindings after unload or reload.